+++ transloadit/request.py
@@ -11,7 +11,7 @@ from .response import as_response
 from . import __version__
 
 
-class Request(object):
+class Request:
     """
     Transloadit tailored HTTP Request object.
 
@@ -23,7 +23,7 @@ class Request(object):
         - transloadit (<transloadit.client.Transloadit>)
     """
 
-    HEADERS = {'Transloadit-Client': 'python-sdk:' + __version__}
+    HEADERS = {"Transloadit-Client": "python-sdk:" + __version__}
 
     def __init__(self, transloadit):
         self.transloadit = transloadit
@@ -39,9 +39,11 @@ class Request(object):
 
         Return an instance of <transloadit.response.Response>
         """
-        return requests.get(self._get_full_url(path),
-                            params=self._to_payload(params),
-                            headers=self.HEADERS)
+        return requests.get(
+            self._get_full_url(path),
+            params=self._to_payload(params),
+            headers=self.HEADERS,
+        )
 
     @as_response
     def post(self, path, data=None, extra_data=None, files=None):
@@ -61,8 +63,9 @@ class Request(object):
         data = self._to_payload(data)
         if extra_data:
             data.update(extra_data)
-        return requests.post(self._get_full_url(path), data=data,
-                             files=files, headers=self.HEADERS)
+        return requests.post(
+            self._get_full_url(path), data=data, files=files, headers=self.HEADERS
+        )
 
     @as_response
     def put(self, path, data=None):
@@ -90,26 +93,27 @@ class Request(object):
         Return an instance of <transloadit.response.Response>
         """
         data = self._to_payload(data)
-        return requests.delete(self._get_full_url(path), data=data, headers=self.HEADERS)
+        return requests.delete(
+            self._get_full_url(path), data=data, headers=self.HEADERS
+        )
 
     def _to_payload(self, data):
         data = copy.deepcopy(data or {})
         expiry = timedelta(seconds=self.transloadit.duration) + datetime.utcnow()
-        data['auth'] = {
-            'key': self.transloadit.auth_key,
-            'expires': expiry.strftime("%Y/%m/%d %H:%M:%S+00:00")
+        data["auth"] = {
+            "key": self.transloadit.auth_key,
+            "expires": expiry.strftime("%Y/%m/%d %H:%M:%S+00:00"),
         }
         json_data = json.dumps(data)
-        return {'params': json_data,
-                'signature': self._sign_data(json_data)}
+        return {"params": json_data, "signature": self._sign_data(json_data)}
 
     def _sign_data(self, message):
-        return hmac.new(b(self.transloadit.auth_secret),
-                        message.encode('utf-8'),
-                        hashlib.sha1).hexdigest()
+        return hmac.new(
+            b(self.transloadit.auth_secret), message.encode("utf-8"), hashlib.sha1
+        ).hexdigest()
 
     def _get_full_url(self, url):
-        if url.startswith(('http://', 'https://')):
+        if url.startswith(("http://", "https://")):
             return url
         else:
             return self.transloadit.service + url