+++ tests/test_request.py
@@ -19,7 +19,7 @@ class RequestTest(unittest.TestCase):
         mock.get(
             url,
             text='{"ok": "it works"}',
-            request_headers={"Transloadit-Client": "python-sdk:0.1.12"},
+            request_headers={"Transloadit-Client": "python-sdk:0.2.0"},
         )
 
         response = self.request.get("/foo")