+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 ### 0.1.12/ 2020-14-12 ###
 
-* Drop the use of requirements.txt for dependencies
-* Update dependency versions
+* Send `transloadit-client` header along with requests
 
 ### 0.1.10/ 2018-27-08 ###