+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-* Drop Python versions before 3.6 as they are unsupported
+* Drop Python versions before 3.7 as they are unsupported
 * Update code to Python 3 syntax
 
 ### 0.1.12/ 2020-14-12 ###